requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
APScheduler==3.10.1
async-timeout==4.0.3
attrs==24.2.0
awscli==1.33.44
backoff==2.2.1
blinker==1.8.2
boto3==1.34.123
botocore==1.34.162
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
exceptiongroup==1.2.2
fastapi==0.110.3
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.7
imageio==2.35.1
importlib_resources==6.4.5
inference==0.21.1
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.1
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.15.1
openai==1.51.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pendulum==3.0.0
piexif==1.1.3
pillow==10.4.0
pluggy==1.5.0
prettytable==3.11.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
PuLP==2.9.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.2
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
pytest==8.3.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
referencing==0.35.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.2
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.6
six==1.16.0
skypilot==0.5.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.39.0
structlog==24.4.0
supervision==0.22.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tifffile==2024.8.30
time-machine==2.15.0
tk==0.1.0
tldextract==5.1.2
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.5
twilio==9.3.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
watchdog==5.0.3
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.13.1
zipp==3.20.2
zxing-cpp==2.2.0
